TTIRI-561 Fix division by zero error
Lines that had a length zero linestring as geometry were causing division by zero errors when trying to project
a point onto them. We don't need to consider these lines as candidates as their adjacent lines are also considered
and so is the point that is the degenerated line. Observe, there is a difference between the length of the line and
the length of its geometry. The first might be positive even though the second is 0, probably due to rounding issues
while converting maps.